pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rlvr-summary"
version = "0.1.0"
description = "Tool-augmented summarizer using RL-VR (Reinforcement Learning from Verifier Rewards)"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "RLVR Summary Team"},
]
keywords = ["reinforcement-learning", "summarization", "nlp", "tool-use"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    # Core DL framework
    "torch>=2.3.0",
    "transformers>=4.42.0",
    
    # RL Framework
    "trl>=0.7.0",
    
    # Fast inference and optimization
    "vllm>=0.2.0",
    "bitsandbytes>=0.43.0",
    "accelerate>=0.20.0",
    
    # NLP utilities
    "spacy>=3.7.0",
    "rank-bm25>=0.2.2",
    
    # Data handling
    "datasets>=2.14.0",
    "pandas>=1.5.0",
    
    # Configuration management
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    
    # Experiment tracking
    "wandb>=0.15.0",
    
    # Additional utilities
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.64.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
"Homepage" = "https://github.com/peara/rlvr-summary"
"Bug Reports" = "https://github.com/peara/rlvr-summary/issues"
"Source" = "https://github.com/peara/rlvr-summary"

[project.scripts]
rlvr-train = "rlvr_summary.cli:train"
rlvr-eval = "rlvr_summary.cli:evaluate"
rlvr-generate = "rlvr_summary.cli:generate"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"rlvr_summary" = ["configs/*.yaml", "configs/**/*.yaml"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["rlvr_summary"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "vllm.*",
    "rank_bm25.*",
    "wandb.*",
    "hydra.*",
    "omegaconf.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/rlvr_summary --cov-report=term-missing --cov-report=html"